+++ libft/lib_fct/lib_str/src/ft_atoi.c
@@ -6,7 +6,7 @@
 /*   By: dbaule <dbaule@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/08 14:07:50 by eslamber          #+#    #+#             */
-/*   Updated: 2023/10/25 10:51:00 by dbaule           ###   ########.fr       */
+/*   Updated: 2023/10/29 17:03:45 by dbaule           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -24,9 +24,9 @@ static int	ft_iswhitespace(int car)
 	return (0);
 }
 
-long long	ft_atoi(const char *str)
+long	ft_atoi(const char *str)
 {
-	long long	res;
+	long		res;
 	size_t		ind;
 	int			minus;
